requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
